Herencia-Kotlin v0.2

+++ herencia-dto/SergioDeIscar-herencia-kotlin/SergioDeIscar-herencia-kotlin/src/main/kotlin/repositories/persona/PersonaExtension.kt
@@ -3,6 +3,6 @@ package repositories.persona
 import models.Persona
 import repositories.CrudRepository
 
-interface PersonaExtension: CrudRepository<Persona, Int>{
+interface PersonaExtension: CrudRepository<Persona, Long>{
     fun getPorcentajePorTipo(): Map<String, Double>
 }
